se crean modulos, falta conexion entre estos
