+++ main.js
@@ -1,0 +1,8 @@
+const mostrarMenu = require('./component/mostrarMenu')
+
+function main() {
+
+    mostrarMenu()
+}
+
+main();
