+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "node App.js",
+    "start": "node main.js",
     "test": "echo \"Error: no test specified\" && exit 1"
 
   },
